feat: admin product media management
- Attach images and video to products using ActiveStorage
- Render media on product show page
- Serve uploaded files via local disk storage

+++ app/controllers/carts_controller.rb
@@ -5,10 +5,7 @@ class CartsController < ApplicationController
 
   def add_item
     product = Product.find(params[:product_id])
-
-    cart_item = current_cart.cart_items.find_or_initialize_by(
-      product: product
-    )
+    cart_item = current_cart.cart_items.find_or_initialize_by(product: product)
 
     if cart_item.new_record?
       cart_item.quantity = 1
@@ -18,64 +15,77 @@ class CartsController < ApplicationController
 
     cart_item.save!
 
+    # Bust the cached cart so next render is fresh
+    @current_cart = nil
+
     respond_to do |format|
-      format.turbo_stream
-      format.html do
-        redirect_back(
-          fallback_location: root_path,
-          notice: "#{product.name} added to your collection"
-        )
-      end
+      format.turbo_stream { render turbo_stream: turbo_stream.replace("cart_content", partial: "carts/cart_content") }
+      format.html { redirect_back(fallback_location: root_path, notice: "#{product.name} added to your collection") }
     end
   end
 
   def increase_quantity
-    cart_item = current_cart.cart_items.find(params[:id])
+    cart_item = current_cart.cart_items.find_by(id: params[:id])
 
-    cart_item.increment!(:quantity)
+    if cart_item.nil?
+      @current_cart = nil
+      respond_to do |format|
+        format.turbo_stream { render turbo_stream: turbo_stream.replace("cart_content", partial: "carts/cart_content") }
+        format.html { redirect_back(fallback_location: cart_path, alert: "Cart updated. Please refresh.") }
+      end
+      return
+    end
+
+    cart_item.update!(quantity: cart_item.quantity + 1)
+    @current_cart = nil  # Bust cache
 
     respond_to do |format|
-      format.turbo_stream
-      format.html do
-        redirect_back(
-          fallback_location: cart_path
-        )
-      end
+      format.turbo_stream { render turbo_stream: turbo_stream.replace("cart_content", partial: "carts/cart_content") }
+      format.html { redirect_back(fallback_location: cart_path) }
     end
   end
 
   def decrease_quantity
-    cart_item = current_cart.cart_items.find(params[:id])
+    cart_item = current_cart.cart_items.find_by(id: params[:id])
+
+    if cart_item.nil?
+      @current_cart = nil
+      respond_to do |format|
+        format.turbo_stream { render turbo_stream: turbo_stream.replace("cart_content", partial: "carts/cart_content") }
+        format.html { redirect_back(fallback_location: cart_path, alert: "Cart updated. Please refresh.") }
+      end
+      return
+    end
 
     if cart_item.quantity > 1
-      cart_item.decrement!(:quantity)
+      cart_item.update!(quantity: cart_item.quantity - 1)
     else
       cart_item.destroy
     end
 
+    @current_cart = nil  # Bust cache
+
     respond_to do |format|
-      format.turbo_stream
-      format.html do
-        redirect_back(
-          fallback_location: cart_path
-        )
-      end
+      format.turbo_stream { render turbo_stream: turbo_stream.replace("cart_content", partial: "carts/cart_content") }
+      format.html { redirect_back(fallback_location: cart_path) }
     end
   end
 
   def remove_item
-    cart_item = current_cart.cart_items.find(params[:id])
+    cart_item = current_cart.cart_items.find_by(id: params[:id])
 
-    cart_item.destroy
+    if cart_item
+      cart_item.destroy
+      notice = "Item removed from cart"
+    else
+      notice = "Item already removed"
+    end
+
+    @current_cart = nil  # Bust cache
 
     respond_to do |format|
-      format.turbo_stream
-      format.html do
-        redirect_back(
-          fallback_location: cart_path,
-          notice: "Item removed from cart"
-        )
-      end
+      format.turbo_stream { render turbo_stream: turbo_stream.replace("cart_content", partial: "carts/cart_content") }
+      format.html { redirect_back(fallback_location: cart_path, notice: notice) }
     end
   end
 end
